refactor: watcher handler
