+++ app/project-manager-shim/src/handler/index.ts
@@ -4,3 +4,4 @@
 
 export * from './filesystem.js'
 export * from './projectService.js'
+export * from './watcher.js'
