+++ app/project-manager-shim/src/handler/watcher.ts
@@ -1,64 +1,129 @@
+import { AssetId, DirectoryId, ProjectId, type GetText } from 'enso-common/src/services/Backend'
+import { HttpClient } from 'enso-common/src/services/HttpClient'
+import { RemoteBackend } from 'enso-common/src/services/RemoteBackend'
+import { getText, resolveDictionary } from 'enso-common/src/text'
 import type * as http from 'node:http'
-import { watch, type Watcher } from '../fs.js'
+import { watch, type Watcher, type WatchOptions } from '../fs.js'
+import * as projectManagement from '../projectManagement.js'
+import { uploadFile } from '../upload.js'
 import { bodyJson } from './http.js'
 
 const HTTP_STATUS_OK = 200
-const HTTP_STATUS_NOT_FOUND = 404
+const HTTP_STATUS_BAD_REQUEST = 400
 const HTTP_STATUS_ERROR = 500
-
-/** Map of directory paths to active watchers */
-const watchers = new Map<string, Watcher>()
 
 /** Check if this is a watcher request */
 export function isWatcherRequest(requestPath: string): boolean {
   return requestPath.startsWith('/api/watcher/')
 }
 
-/** Handle watcher requests */
+/** Handle watcher requests. */
 export async function handleWatcherRequest(
   request: http.IncomingMessage,
   response: http.ServerResponse,
   requestPath: string,
   headers: Record<string, string>,
-  callback: () => Promise<void>,
+  watchers: Map<AssetId, Watcher>,
+  options: Pick<WatchOptions, 'delay' | 'timeout'>,
 ): Promise<void> {
+  const url = new URL(request.url ?? '', 'https://apishim.local')
   switch (`${request.method} ${requestPath}`) {
-    case 'POST /api/watcher/watch': {
-      interface Body {
-        readonly directory: string
-        readonly delay: number
-        readonly timeout: number
+    case 'POST /api/watcher/start': {
+      const projectDir = url.searchParams.get('directory')
+      if (projectDir == null) {
+        response
+          .writeHead(HTTP_STATUS_BAD_REQUEST, headers)
+          .end('Request is missing search parameter `directory`.')
+        break
       }
+      const assetIdString = url.searchParams.get('assetId')
+      if (assetIdString == null) {
+        response
+          .writeHead(HTTP_STATUS_BAD_REQUEST, headers)
+          .end('Request is missing search parameter `assetId`.')
+        break
+      }
+      const parentDirectoryIdString = url.searchParams.get('parentDirectoryId')
+      if (parentDirectoryIdString == null) {
+        response
+          .writeHead(HTTP_STATUS_BAD_REQUEST, headers)
+          .end('Request is missing search parameter `parentDirectoryId`.')
+        break
+      }
+      const baseUrl = url.searchParams.get('baseUrl')
+      if (baseUrl == null) {
+        response
+          .writeHead(HTTP_STATUS_BAD_REQUEST, headers)
+          .end('Request is missing search parameter `baseUrl`.')
+        break
+      }
+      const assetId = ProjectId(assetIdString)
+
+      console.log('DEBUG start watching project', assetId)
       try {
-        const { directory, delay, timeout } = await bodyJson<Body>(request)
-        const existingWatcher = watchers.get(directory)
-        if (existingWatcher) {
-          await existingWatcher.close()
+        const defaultHeaders = await bodyJson<Record<string, string>>(request)
+        const client = new HttpClient(defaultHeaders)
+        const downloader = () => {}
+        const dictionary = resolveDictionary()
+        const backendGetText: GetText = function (key, ...replacements) {
+          return getText(dictionary, key, ...replacements)
         }
-        const watcher = watch({ directory, delay, timeout, callback })
-        watchers.set(directory, watcher)
+        const backend = new RemoteBackend(backendGetText, client, downloader, new URL(baseUrl))
+        const fileName = 'project_root.enso-project'
+        const uploadParams = {
+          fileId: assetId,
+          fileName,
+          parentDirectoryId: parentDirectoryIdString as DirectoryId,
+        }
+        const watcher = watch({
+          directory: projectDir,
+          delay: options.delay,
+          timeout: options.timeout,
+          callback: async () => {
+            console.log('DEBUG UPLOADING STARTED')
+            const responseBody = await projectManagement.createBundle(projectDir)
+            const file = new File([responseBody.buffer as ArrayBuffer], fileName)
+            await uploadFile(backend, uploadParams, file)
+            console.log('DEBUG UPLOADING FINISHED')
+          },
+        })
+
+        const existingWatcher = watchers.get(assetId)
+        if (existingWatcher) {
+          await existingWatcher.close().catch((err) => {
+            console.error(`Failed to stop project watcher ${assetId}`, err)
+          })
+        }
+        watchers.set(assetId, watcher)
         response.writeHead(HTTP_STATUS_OK, headers).end()
-      } catch (error) {
-        response.writeHead(HTTP_STATUS_ERROR, headers).end(JSON.stringify({ error }))
+      } catch (err) {
+        console.error(`Failed to start project watcher ${assetId}`, err)
+        response.writeHead(HTTP_STATUS_ERROR, headers).end()
       }
       break
     }
-    case 'POST /api/watcher/close': {
-      interface Body {
-        readonly directory: string
+    case 'POST /api/watcher/stop': {
+      const assetIdString = url.searchParams.get('assetId')
+      if (assetIdString == null) {
+        response
+          .writeHead(HTTP_STATUS_BAD_REQUEST, headers)
+          .end('Request is missing search parameter `assetId`.')
+        break
       }
-      try {
-        const { directory } = await bodyJson<Body>(request)
-        const watcher = watchers.get(directory)
-        if (watcher) {
-          const isUnsaved = await watcher.close()
-          watchers.delete(directory)
-          response.writeHead(HTTP_STATUS_OK, headers).end(JSON.stringify({ isUnsaved }))
-        } else {
-          response.writeHead(HTTP_STATUS_NOT_FOUND, headers).end()
-        }
-      } catch (error) {
-        response.writeHead(HTTP_STATUS_ERROR, headers).end(JSON.stringify({ error }))
+      const assetId = ProjectId(assetIdString)
+      console.log('DEBUG stop watching project', assetId)
+      const watcher = watchers.get(assetId)
+      if (watcher) {
+        await watcher
+          .close()
+          .catch((err) => {
+            console.error(`Failed to stop project watcher ${assetId}`, err)
+          })
+          .finally(() => {
+            response.writeHead(HTTP_STATUS_OK, headers).end()
+          })
+      } else {
+        response.writeHead(HTTP_STATUS_OK, headers).end()
       }
       break
     }
